@@ -63,46 +63,56 @@ let commandConfig = {
 
             return data;
         },
-        clearRow: '[2K',  // 在cmd中用于清除当前行并输入后面内容，即命令行的进度条。未知正确字符串，目前只能以乱码形式显示在ide中
-        endRow: '[1B',
-
-
-        // 命令是否执行结束处理
-        /**
-         * @return {string}
-         */
-        ShellEOFCommand(eof){
-            return `echo "${eof}" \n`;
+        openShell(command){
+            let environment = proxyConfig.shellName;
+    
+            if (environment === 'cmd'){
+                return `start cmd /k ${command}`;
+            }else if(environment === 'git-bash'){
+                // todo 待改为git的bash，记得有部分功能有用
+                return `start ${dockerConfig.bashCommandPath} -s -c '${command} ;read'`;
+            }else if(environment === 'powershell'){
+                command = `echo '${command.replace(/(^`)'/g, "$1''")}';${command}`; // 输出执行命令
+                command = command.replace(/([^`])"/g, "$1\"\"\"\"\"\""); // 转双引号需要三对双引号
+                // return `start powershell -NoExit ${docker_compose} logs ${logsArgs} ${getContainerName()}`; // cmd下打开powershell
+                return `start powershell -ArgumentList "-NoExit ${command}"`; // powershell打开powershell
+            }
         },
 
-        enter: [ // 留意替换顺序
-            "\n", // todo 换行
-            "\r\n", // todo \r\n\r\n命令结束
-        ],
-        /**
-         * 不是回车符
-         *
-         * @param str
-         * @returns {boolean}
-         */
-        ifNotEnter(str){
-            return !(str.length === 2 && this.enter.includes(str))
-        },
 
-        /**
-         * 换行切换数组
-         *
-         * @param str
-         * @returns {Array}
-         */
-        lineArr(str){
-            return str.replace("\r\n\r\n", "").split(/\r?\n/) || [];
-        },
+    },
 
-        getNotEnter(str){
-            return str.trim()
+
+
+    
+    linux:{
+        whereCmd(command){
+            return `which ${command}`;
         }
     },
+    mac:{
+        whereCmd(command){
+            return `which ${command}`;
+        },
+        openShell(command){
+            return config.macTerminalExecPath + ' ' + command;
+        },
+    },
+
+
+
+
+
+
+
+    get headArr(){
+        return [];
+    },
+
+    
+    clearRow: '[2K',  // 在cmd中用于清除当前行并输入后面内容，即命令行的进度条。未知正确字符串，目前只能以乱码形式显示在ide中
+    endRow: '[1B',
+
 
     replaceEnter(str, replace, num = null){
         num = (num === null ? this.enter.length : num) - 1;
@@ -110,31 +120,63 @@ let commandConfig = {
         return this.enter[num] ? this.replaceEnter(str.replace(this.enter[num], replace), replace, num) : str;
     },
 
-    openShell(command){
-        let environment = proxyConfig.shellName;
+    /**
+     * 命令是否执行结束处理
+     * 
+     * @return {string}
+     */
+    ShellEOFCommand(eof){
+        return `echo "${eof}" \n`;
+    },
 
-        if (environment === 'cmd'){
-            return `start cmd /k ${command}`;
-        }else if(environment === 'git-bash'){
-            // todo 待改为git的bash，记得有部分功能有用
-            return `start ${dockerConfig.bashCommandPath} -s -c '${command} ;read'`;
-        }else if(environment === 'powershell'){
-            command = `echo '${command.replace(/(^`)'/g, "$1''")}';${command}`; // 输出执行命令
-            command = command.replace(/([^`])"/g, "$1\"\"\"\"\"\""); // 转双引号需要三对双引号
-            // return `start powershell -NoExit ${docker_compose} logs ${logsArgs} ${getContainerName()}`; // cmd下打开powershell
-            return `start powershell -ArgumentList "-NoExit ${command}"`; // powershell打开powershell
-        }
+
+    /**
+     * 不是回车符
+     *
+     * @param str
+     * @returns {boolean}
+     */
+    ifNotEnter(str){
+        return !(str.length === 2 && this.enter.includes(str))
     },
-    linux:{
-        whereCmd(command){
-            return `which ${command}`;
-        },
+
+    /**
+     * 换行切换数组
+     *
+     * @param str
+     * @returns {Array}
+     */
+    lineArr(str){
+        return str.replace("\r\n\r\n", "").split(/\r?\n/) || [];
     },
-    mac:{
-        whereCmd(command){
-            return `which ${command}`;
-        },
+
+    getNotEnter(str){
+        return str.trim()
     },
+
+
+    get shellName(){
+        return 'bash';
+    },
+
+    get execShellName(){
+        return 'bash';
+    },
+    enter: [ // 留意替换顺序
+        "\n", // todo 换行
+        "\r\n", // todo \r\n\r\n命令结束
+    ],
+
+    get headArr(){
+        return [];
+    },
+    openShell(command){
+
+    },
+
+
+
+
 
     /**
      * 查看command命令所在路径
@@ -192,14 +234,18 @@ let commandConfig = {
 
 proxyConfig = new Proxy(commandConfig, {
     get:(target, key) => {
-        if (!Reflect.has(target, key)){ // 顶层不存在即读取系统配置层
-            if (isWin) { // todo 可能要改为win
-                target = target['windows'];
-            } else if(isLinux){
-                target = target['linux'];
-            } else if (isMac){
-                target = target['mac'];
-            }
+        let data;
+        if (isWin) { // todo 可能要改为win
+            data = target['windows'];
+        } else if(isLinux){
+            data = target['linux'];
+        } else if (isMac){
+            data = target['mac'];
+        }
+
+
+        if (Reflect.has(data, key)){ // 系统配置存在
+            target = data;
         }
 
         return Reflect.has(target, key) ? target[key] : false;
