@@ -1,12 +1,16 @@
 // webpack.config.js
 const config = require(__dirname + '/../config/config.js');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const VueLoaderPlugin = require('vue-loader/lib/plugin');
 const webpack = require('webpack');
 const Path = require('path');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+
 global.config = config; // 解决nodeStorage
 
-
-module.exports = {
+let webpackConfig = {};
+webpackConfig = {
+    mode: 'development', // production, development
     devtool: '#cheap-module-eval-source-map', // 用生成环境（production supported）为no的可以不显示黄色警告
     // devtool: '#source-map', // 最好的错误提示
     target: 'electron-renderer', // 可以直接require('electron')，而不必window.require('electron')
@@ -20,54 +24,75 @@ module.exports = {
         filename: '[name].js'
     },
     module: {
-        preLoaders: [],
         // `loaders` is an array of loaders to use.
         // here we are only configuring vue-loader
-        loaders: [
+        rules: [
             {
                 test: /\.vue$/, // a regex for matching all files that end in `.vue`
-                loader: 'vue-loader',   // loader to use for matched files
-                exclude: /node_modules/,
+                use: 'vue-loader',   // loader to use for matched files
             },
             {
-                "test": /\.css?$/,
-                "loader": "style!css",
-                exclude: /node_modules/,
+                "test": /\.css$/,
+                "use": [
+                    MiniCssExtractPlugin.loader,
+                    // "vue-style-loader",
+                    "css-loader"
+                ],
             },
             { // 不知为何<% %>解析不了
                 "test": /\.html$/,
-                "loader": "vue-html-loader",
-                exclude: /node_modules/,
+                "use": "vue-html-loader",
             },
             {
                 test: /\.js$/,
-                loader:'babel-loader',
-                exclude: /node_modules/
+                use: {
+                    loader: 'babel-loader',
+                    options: { // 使用babel，省去.babelrc
+                        // enable stage 0 babel transforms.
+                        presets: ['env', 'stage-0'],
+                        plugins: ['transform-runtime']
+                    }
+                },
+                exclude: /(node_modules|bower_components)/,
             }
         ]
     },
     resolve:{
-        root : config.srcPath,
-        fallback:config.srcPath + '/node_modules',
-        extensions: ['', '.js', '.vue', '.json', '.css'],
+        modules: [
+            config.rootPath + '/node_modules',
+            config.srcPath + '/node_modules'
+        ],
+        extensions: ['.js', '.vue', '.json', '.css'],
         alias : {
             appMain:"/temp/appMain.js",
-            'vue$': 'vue/dist/vue.js',
+            'vue$': 'vue/dist/vue.min.js', // electron/node_modules/vue/dist/README.md
+            'vue-router$': 'vue-router/dist/vue-router.min.js',
+            appDir: config.srcPath + '/app', // unused
         }
     },
-    resolveLoader: {
-        root: Path.join(config.rootPath, 'node_modules')
-    },
-    babel: { // 使用babel，省去.babelrc
-        // enable stage 0 babel transforms.
-        presets: ['es2015', 'stage-0'],
-        plugins: ['transform-runtime']
-    },
-    devServer:{ // 木有用
+    // resolveLoader: {
+    //     modules: [
+    //         Path.join(config.rootPath, 'node_modules')
+    //     ],
+    //     extensions: ['.js', '.vue', '.json', '.css'],
+    // },
+    devServer:{ // 木有用，直接在server.js中另写了
         hot: true,
         inline: true,
     },
     plugins:[
+        new MiniCssExtractPlugin({ // 将css都放到appMain.css中
+            // Options similar to the same options in webpackOptions.output
+            // both options are optional
+            filename: "[name].css",
+            chunkFilename: "[id].css"
+        }),
+        new webpack.DefinePlugin({
+            'process.env': {
+                NODE_ENV: '"'+webpackConfig.mode+'"'
+            }
+        }),
+        new VueLoaderPlugin(),
         new HtmlWebpackPlugin({ // 让index.html内可自动处理编译后的appMain.js
             filename: config.indexHtmlName,
             template: config.indexHtmlPathSource,
@@ -81,3 +106,18 @@ module.exports = {
         // }),
     ]
 };
+
+module.exports = webpackConfig;
+
+
+if (process.env.NODE_ENV === 'production') {
+    module.exports.devtool = '#source-map';
+    // http://vue-loader.vuejs.org/en/workflow/production.html
+    module.exports.plugins = (module.exports.plugins || []).concat([
+        // new webpack.optimize.UglifyJsPlugin({
+        //     compress: {
+        //         warnings: false
+        //     }
+        // })
+    ])
+}
